refactor article

+++ src/main/java/edu/tongji/article/ArticleRepository.java
@@ -1,32 +1,23 @@
 package edu.tongji.article;
 
 import edu.tongji.account.Account;
-import org.pegdown.PegDownProcessor;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 import javax.persistence.PersistenceException;
-import javax.persistence.Transient;
 import java.util.List;
 
 @Repository
-//@Transactional(readOnly = true)
+@Transactional
 public class ArticleRepository {
 
     @PersistenceContext
     private EntityManager entityManager;
 
-    @Autowired
-    @Transient
-    private PegDownProcessor pegDownProcessor;
-
     @Transactional
     public Article save(Article article) {
-        article.setHtml(pegDownProcessor.markdownToHtml(article.getMarkdown()));
-        article.setUrl(article.getTitle());
         entityManager.persist(article);
         return article;
     }
@@ -55,16 +46,20 @@ public class ArticleRepository {
     @Transactional
     public Boolean delete(Account account, Long id) {
         Article article = getArticle(account, id);
-        entityManager.remove(article);
+        if (article != null) {
+            entityManager.remove(article);
+        }
         return true;
     }
 
     @Transactional
     public Article update(Account account, Long id, String markdown, String title) {
         Article article = getArticle(account, id);
-        article.setMarkdown(markdown);
-        article.setTitle(title);
-        entityManager.merge(article);
+        if (article != null) {
+            article.setMarkdown(markdown);
+            article.setTitle(title);
+            entityManager.merge(article);
+        }
         return article;
     }
 }
